backend/requirements.txt
opencv-python
numpy
matplotlib
scikit-learn
pytesseract
Pillow
uvicorn
fastapi
python-multipart
python-dotenv
pyyaml
pydantic
google-genai
